Player and System implementation P1
Implemented the player and all the systems up until this point. However, I am going to revise the AccuseSystem Script to automatically check and fill out how many npc's there are instead of me manually declaring there are 4. If i did this would also create an error in the dev area unless I have the same amount of NPCs in the across these scenes